+++ tfe/resource_tfe_project_variable_set_test.go
@@ -16,7 +16,6 @@ import (
 )
 
 func TestAccTFEProjectVariableSet_basic(t *testing.T) {
-	skipUnlessBeta(t)
 	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
 
 	tfeClient, err := getClientUsingEnv()