+++ tfe/resource_tfe_team_test.go
@@ -77,6 +77,8 @@ func TestAccTFETeam_full(t *testing.T) {
 						"tfe_team.foobar", "organization_access.0.read_projects", "true"),
 					resource.TestCheckResourceAttr(
 						"tfe_team.foobar", "organization_access.0.read_workspaces", "true"),
+					resource.TestCheckResourceAttr(
+						"tfe_team.foobar", "organization_access.0.manage_membership", "true"),
 				),
 			},
 		},
@@ -122,6 +124,8 @@ func TestAccTFETeam_full_update(t *testing.T) {
 						"tfe_team.foobar", "organization_access.0.manage_projects", "true"),
 					resource.TestCheckResourceAttr(
 						"tfe_team.foobar", "organization_access.0.read_workspaces", "true"),
+					resource.TestCheckResourceAttr(
+						"tfe_team.foobar", "organization_access.0.manage_membership", "true"),
 				),
 			},
 			{
@@ -154,6 +158,8 @@ func TestAccTFETeam_full_update(t *testing.T) {
 						"tfe_team.foobar", "organization_access.0.read_projects", "false"),
 					resource.TestCheckResourceAttr(
 						"tfe_team.foobar", "sso_team_id", "changed-sso-id"),
+					resource.TestCheckResourceAttr(
+						"tfe_team.foobar", "organization_access.0.manage_membership", "false"),
 				),
 			},
 			{
@@ -187,6 +193,8 @@ func TestAccTFETeam_full_update(t *testing.T) {
 						"tfe_team.foobar", "organization_access.0.read_projects", "false"),
 					resource.TestCheckResourceAttr(
 						"tfe_team.foobar", "sso_team_id", ""),
+					resource.TestCheckResourceAttr(
+						"tfe_team.foobar", "organization_access.0.manage_membership", "false"),
 				),
 			},
 		},
@@ -436,6 +444,9 @@ func testAccCheckTFETeamAttributes_full(
 		if !team.OrganizationAccess.ManageProjects {
 			return fmt.Errorf("OrganizationAccess.ManageProjects should be true")
 		}
+		if !team.OrganizationAccess.ManageMembership {
+			return fmt.Errorf("OrganizationAccess.ManageMembership should be true")
+		}
 		if team.SSOTeamID != "team-test-sso-id" {
 			return fmt.Errorf("Bad SSO Team ID: %s", team.SSOTeamID)
 		}
@@ -469,6 +480,9 @@ func testAccCheckTFETeamAttributes_full_update(
 		}
 		if team.OrganizationAccess.ManageProjects {
 			return fmt.Errorf("OrganizationAccess.ManageProjects should be false")
+		}
+		if team.OrganizationAccess.ManageMembership {
+			return fmt.Errorf("OrganizationAccess.ManageMembership should be false")
 		}
 
 		if team.SSOTeamID != "changed-sso-id" {
@@ -537,6 +551,7 @@ resource "tfe_team" "foobar" {
 	manage_projects = true
 	read_workspaces = true
 	read_projects = true
+	manage_membership = true
   }
   sso_team_id = "team-test-sso-id"
 }`, rInt)
@@ -566,6 +581,7 @@ resource "tfe_team" "foobar" {
 	manage_projects = false
 	read_projects = false
 	read_workspaces = false
+	manage_membership = false
   }
 
   sso_team_id = "changed-sso-id"